Revert "makefile outputs a gfm readme now"
i changed the yaml header to output both a html document and github
flavored md document. but i also had set to keep_md true with the html
document first. so if you ran make you'd overright the keeped md if that
makes sense, but if you just ran knit in rstudio for the html then you'd
get the html and the md document (regular flavor) and you wouldn't over
write anything. i'm undoing this because it seems unneccessarily complicat-
ed and it created another copy of all the figures and i still couldn't
get a toc in the md. the output with keep_md is decent and renders fine
on github it just thinks the yaml header is a table which is fine and there
is no toc. but i'll see if i can get the toc easily.

This reverts commit dfbb8ee9575214b98fb71a312afaa97a5ff25038.